Auth, design fixes
Changed auth handling to the server-side, design changes and added some of our future expected games to the explore page. Merged the login page with landing for a more seamless user experienc. Additionally conditional rendering was added to some parts of the landing page so that you only get the sign-in options if you aren't logged etc.