New parts of the app
The settings page now lets the user change their username. The lobby page, is still WIP but im still impressed of what I've been able to achieve with it so far. All considered, we now already have a solid foundation for the lobby system when we get the actual games in.